requirements.txt
absl-py==0.12.0
astor==0.8.1
backcall==0.2.0
certifi==2016.2.28
click==7.1.2
dataclasses==0.8
decorator==5.0.5
Flask==1.1.2
gast==0.4.0
gensim==3.6.0
grpcio==1.36.1
importlib-metadata==3.10.0
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
jieba==0.42.1
Jinja2==2.11.3
joblib==1.0.1
Markdown==3.3.4
MarkupSafe==1.1.1
numpy==1.14.5
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
protobuf==3.15.6
ptyprocess==0.7.0
Pygments==2.8.1
scikit-learn==0.24.1
scipy==1.5.4
six==1.15.0
sklearn==0.0
smart-open==5.0.0
tensorboard==1.10.0
tensorflow-gpu @ file:///home/ps/disk_sdb/yyr/tensorflow_gpu-1.10.1-cp36-cp36m-manylinux1_x86_64.whl
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.59.0
traitlets==4.3.3
typing-extensions==3.7.4.3
ujson==4.0.2
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==3.4.1
